Fix a bunch of minor MSVC warnings (integral types stuff)
Former-commit-id: 96342d058a7b1da95d2c81f5d1b933632656e6aa

+++ modules/SofaBaseMechanics/DiagonalMass.inl
@@ -432,19 +432,19 @@ void DiagonalMass<DataTypes, MassType>::addMDx(const core::MechanicalParams* /*m
     helper::WriteAccessor< DataVecDeriv > _res = res;
     helper::ReadAccessor< DataVecDeriv > _dx = dx;
 
-    unsigned int n = masses.size();
+    size_t n = masses.size();
     if (_dx.size() < n) n = _dx.size();
     if (_res.size() < n) n = _res.size();
     if (factor == 1.0)
     {
-        for (unsigned int i=0; i<n; i++)
+        for (size_t i=0; i<n; i++)
         {
             _res[i] += _dx[i] * masses[i];
         }
     }
     else
     {
-        for (unsigned int i=0; i<n; i++)
+        for (size_t i=0; i<n; i++)
         {
             _res[i] += (_dx[i] * masses[i]) * (Real)factor;
         }
@@ -751,8 +751,8 @@ void DiagonalMass<DataTypes, MassType>::init()
     if (this->mstate && f_mass.getValue().size() > 0 && f_mass.getValue().size() < (unsigned)this->mstate->getSize())
     {
         MassVector &masses= *f_mass.beginEdit();
-        unsigned int i = masses.size()-1;
-        unsigned int n = (unsigned)this->mstate->getSize();
+        size_t i = masses.size()-1;
+        size_t n = (size_t)this->mstate->getSize();
         masses.reserve(n);
         while (masses.size() < n)
             masses.push_back(masses[i]);
